refactor structures
