+++ config/jms_serializer/Model.CheckClose.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <serializer>
-    <class name="OrangeData\Model\CheckClose" exclusion-policy="NONE">
-        <property name="payments" type="array<OrangeData\Model\Payment>"/>
+    <class name="OrangeData\Structure\CheckClose" exclusion-policy="NONE">
+        <property name="payments" type="array<OrangeData\Structure\Payment>"/>
         <property name="taxationSystem" type="int" serialized-name="taxationSystem"/>
     </class>
 </serializer>
