+++ config/jms_serializer/Model.Payment.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <serializer>
-    <class name="OrangeData\Model\Payment" exclusion-policy="NONE">
+    <class name="OrangeData\Structure\Payment" exclusion-policy="NONE">
         <property name="type" type="int"/>
         <property name="amount" type="float"/>
     </class>
