+++ src/Body/BodyInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * @author Dmitry Anikeev <anikeev.dmitry@outlook.com>
+ */
+
+namespace OrangeData\Body;
+
+interface BodyInterface
+{
+    public function asString(): string;
+}
