--- src/Client/SignInterface.php
@@ -1,10 +1,0 @@
-<?php declare(strict_types=1);
-
-
-namespace OrangeData\Client;
-
-
-interface SignInterface
-{
-    public function toString(): string;
-}
