+++ src/Sign/SignInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+
+namespace OrangeData\Sign;
+
+interface SignInterface
+{
+    public function asString(): string;
+}
