+++ src/Structure/AgentInfo.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
  * @author Anikeev Dmitry <dm.anikeev@gmail.com>
  */
 
-namespace OrangeData\Model;
+namespace OrangeData\Structure;
 
 use JsonSerializable;
 
 /**
  * Class AgentInfo
- * @package OrangeData\Model
+ * @package OrangeData\Structure
  */
 class AgentInfo implements JsonSerializable
 {
