+++ src/Structure/SupplierInfo.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
  * @author Anikeev Dmitry <dm.anikeev@gmail.com>
  */
 
-namespace OrangeData\Model;
+namespace OrangeData\Structure;
 
 use JsonSerializable;
